+++ Rebus/Pipeline/IMessageContext.cs
@@ -1,4 +1,5 @@
-﻿using Rebus.Messages;
+﻿using System.Collections.Generic;
+using Rebus.Messages;
 using Rebus.Transport;
 
 namespace Rebus.Pipeline
@@ -30,5 +31,11 @@ namespace Rebus.Pipeline
         /// Gets the <see cref="Message"/> model for the message currently being handled.
         /// </summary>
         Message Message { get; }
+
+        /// <summary>
+        /// Gets the headers dictionary of the incoming message (same as accessing <see cref="IMessageContext.TransportMessage.Headers"/>
+        /// or <see cref="IMessageContext.Message.Headers"/> if the message has been properly deserialized)
+        /// </summary>
+        Dictionary<string,string> Headers { get; }
     }
 }